requirements.txt
asgiref==3.7.2
Brotli==1.0.9
cached-property==1.5.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.2
django-autoslug==1.9.9
django-brotli==0.2.1
django-cities-light==3.9.2
django-cors-headers==4.1.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
progressbar2==4.2.0
psycopg2==2.9.6
PyJWT==1.7.1
python-decouple==3.8
python-stdnum==1.18
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
wt==2.1.1
